Minor fixes (addressed comments).
